pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sd-metaviewer"
version = "1.0.0"
description = "A lightweight, native application to view and extract metadata from AI-generated images"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "SD MetaViewer"}
]
keywords = ["stable-diffusion", "metadata", "viewer", "comfyui", "automatic1111", "novelai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Viewers",
]
dependencies = [
    "Pillow>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0.0",
    "windnd>=1.0.0",
]

[project.scripts]
sd-metaviewer = "src.app:main"

[project.gui-scripts]
sd-metaviewer-gui = "src.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
